# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Imagesystems",
    "description": "Imagesystems \u2014 hotel & real estate image enhancement, generation, and batch pipelines",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/images/upload": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Upload Images",
        "operationId": "upload_images_api_images_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_images_api_images_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImageUploadResponse"
                  },
                  "type": "array",
                  "title": "Response Upload Images Api Images Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/images/{image_id}/enhance": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Enhance Image",
        "description": "Start an AI enhancement job.",
        "operationId": "enhance_image_api_images__image_id__enhance_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnhancementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{image_id}/upscale": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Upscale Image",
        "description": "Start an upscaling job.",
        "operationId": "upscale_image_api_images__image_id__upscale_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpscaleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{image_id}/process": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Process Full Pipeline",
        "description": "Start full pipeline: enhance + upscale.",
        "operationId": "process_full_pipeline_api_images__image_id__process_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullPipelineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{image_id}/local-improve": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Upload Local Improve Version",
        "description": "Persist a browser-processed result (Improve engine) as a new final version \u2014 no external API keys.",
        "operationId": "upload_local_improve_version_api_images__image_id__local_improve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_local_improve_version_api_images__image_id__local_improve_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{image_id}": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Get Image",
        "operationId": "get_image_api_images__image_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "images"
        ],
        "summary": "Delete Image",
        "operationId": "delete_image_api_images__image_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{image_id}/download": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Download Image",
        "operationId": "download_image_api_images__image_id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Version ID to download. If empty, downloads original.",
              "title": "Version"
            },
            "description": "Version ID to download. If empty, downloads original."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{image_id}/suggest-filename": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Suggest Export Filename",
        "description": "Vision model suggests a short kebab-case filename stem (requires OpenAI or Gemini API key).",
        "operationId": "suggest_export_filename_api_images__image_id__suggest_filename_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestFilenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestFilenameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "List Images",
        "operationId": "list_images_api_images_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageDetailResponse"
                  },
                  "title": "Response List Images Api Images Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/estimate-cost": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Estimate Cost",
        "description": "Estimate processing cost before starting.",
        "operationId": "estimate_cost_api_images_estimate_cost_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostEstimateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_images_api_images_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_images_api_images_upload_post"
      },
      "Body_upload_local_improve_version_api_images__image_id__local_improve_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_local_improve_version_api_images__image_id__local_improve_post"
      },
...
      "EnhancementRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "openai"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "gpt-image-1"
          },
          "lighting": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lighting"
          },
          "quality_preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Preset"
          },
          "perspective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perspective"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type",
            "default": "general"
          },
          "custom_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100000
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Prompt"
          },
          "output_format": {
            "type": "string",
            "title": "Output Format",
            "default": "png"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "high"
          },
          "auto_rotation_rad": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Rotation Rad",
            "description": "Roll in radians from browser Sobel estimator; used with auto perspective + cloud plate."
          },
          "improve_input_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improve Input Version Id",
            "description": "Required when provider is openai or gemini: ImageVersion id from POST .../local-improve (browser Improve output). Cloud models always enhance this raster, never the raw upload."
          }
        },
        "type": "object",
        "title": "EnhancementRequest"
      },
...
      "FullPipelineRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "openai"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "gpt-image-1"
          },
          "lighting": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lighting",
            "default": "bright"
          },
          "quality_preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Preset",
            "default": "full_enhance"
          },
          "perspective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perspective"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type",
            "default": "general"
          },
          "custom_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100000
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Prompt"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "high"
          },
          "auto_rotation_rad": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Rotation Rad",
            "description": "Roll in radians for auto perspective when using cloud enhance."
          },
          "improve_input_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improve Input Version Id",
            "description": "Required for OpenAI/Gemini full pipeline: id of an ImageVersion with provider=improve (browser Improve output). Cloud models enhance this raster, not the raw upload."
          },
          "scale_factor": {
            "type": "integer",
            "title": "Scale Factor",
            "default": 2
          },
          "target_resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Resolution",
            "default": "4k"
          },
          "output_format": {
            "type": "string",
            "title": "Output Format",
            "default": "png"
          }
        },
        "type": "object",
        "title": "FullPipelineRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ImageVersionResponse"
            },
            "type": "array",
            "title": "Versions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "created_at"
        ],
        "title": "ImageDetailResponse"
      },
...
      "ImageUploadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ImageVersionResponse"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "created_at"
        ],
        "title": "ImageUploadResponse"
      },
...
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_pct": {
            "type": "integer",
            "title": "Progress Pct"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "result_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Version Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_id",
          "job_type",
          "status",
          "progress_pct",
          "created_at"
        ],
        "title": "JobResponse"
      },
...
      "SuggestFilenameRequest": {
        "properties": {
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "Version ID; omit for original."
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "'openai' or 'gemini' \u2014 must match a saved API key. Gemini 2.5 Flash-Lite is default for low-cost naming.",
            "default": "gemini"
          }
        },
        "type": "object",
        "title": "SuggestFilenameRequest",
        "description": "Which image bytes to analyze (original vs a version)."
      },
      "SuggestFilenameResponse": {
        "properties": {
          "basename": {
            "type": "string",
            "title": "Basename"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          },
          "cost_note": {
            "type": "string",
            "title": "Cost Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "basename"
        ],
        "title": "SuggestFilenameResponse"
      },
      "UpscaleRequest": {
        "properties": {
          "scale_factor": {
            "type": "integer",
            "title": "Scale Factor",
            "default": 2
          },
          "target_resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Resolution"
          },
          "output_format": {
            "type": "string",
            "title": "Output Format",
            "default": "png"
          }
        },
        "type": "object",
        "title": "UpscaleRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}